rating

+++ RoseCoffeeIOS/RoseCoffeeIOS/AppDelegate.swift
@@ -20,7 +20,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     var email: String?
     var phoneNum: String?
     var isDelivery: Bool?
-    var balance: Float?
+    var balance: Double?
 
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplicationLaunchOptionsKey: Any]?) -> Bool {
         // Override point for customization after application launch.
